requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0
av>=8.0.0
scikit-image>=0.18.0
imageio>=2.9.0